refactor(pdf): isolate preformatted block evidence
